refactor(llmadmin): Prevent duplicate API calls and extract StatusBadges components
