+++ src/app/[locale]/admin/llm-models/components/LlmModelFilter.tsx
@@ -6,8 +6,8 @@ import { Label } from '@/components/ui/label';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent } from '@/components/ui/card';
-import { GetLlmModelsQuery } from '@/types/llmModel';
-import { LlmProvider, LlmProvidersResponse } from '@/types/llmProvider';
+import { GetLlmModelsQuery } from '@/types/llm-admin/llmModel';
+import { LlmProvider, LlmProvidersResponse } from '@/types/llm-admin/llmProvider';
 import { X } from 'lucide-react';
 import { getProviderColor } from '@/utils/providerColors';
 import { getRequest } from '@/api/api';
